[MOV] Context to dedicated file
No code changes: just move and imports update.
