+++ server/src/core/mod.rs
@@ -4,6 +4,7 @@ pub mod csv_validation;
 pub mod data_hooks;
 pub mod diagnostics;
 pub mod evaluation;
+pub mod evaluation_context;
 pub mod entry_point;
 pub mod file_mgr;
 pub mod import_resolver;
